requirements.txt
transformers
pillow
torch
deep_translator